requirement.txt
colorama
pyyaml
columnar
keyboard
pynput
inquirer